backend/requirements.txt
# Api Section
Uvicorn
FastAPI

#for database
SQLAlchemy
databases[sqlite]

#for pydantic email validation
pydantic[email]

#for password hashing
passlib[bcrypt]

# For Dot Env
python-dotenv

# For Request Validation
requests

# For JWT
python-jose[cryptography]
